FROMLIST: mmc: mtk-sd: add support to disable single burst
Add support to disable 'single' burst type if the bus type is AXI.
Since the AMBA within some of the legacy and new designed MSDC IP
is AXI, this switch is necessary.

The burst type is not IC-specific, but host-specific. So we use a
devicetree property 'mediatek,disable-single-burst' to switch burst
type for specific MSDC host.

Link: https://patchwork.kernel.org/project/linux-mediatek/patch/20250306085028.5024-3-axe.yang@mediatek.com/

Change-Id: I165dc8f8d2298cd6883b3701537872df48250884
Signed-off-by: Axe Yang <axe.yang@mediatek.com>