Revert "GENIO: drm/mediatek: enable renderD for mediatek-drm"
mtk_drm_drv was not supposed to act as an render node,
because "Render nodes solely serve render clients, that is,
no modesetting or privileged ioctls can be issued on render nodes."[1]

The DRIVER_RENDER flag was added to allow applications to allocate DRM
DMABUF when using proprietary MALI driver. However, acting as a render
node breaks upstream graphics stack, as it interferes the render node
provided by the Panfrost GPU driver.

So remove this patch from kernel source tree and only conditionally apply
this patch when proprietary MALI drivers are being used.

[1] https://www.kernel.org/doc/html/v6.6/gpu/drm-uapi.html#render-nodes

Change-Id: I3abc9d2800c70596d6fb6b33e1643bd6deb5a942
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>