GENIO: drm: mediatek: add display dynamic route for dvo2
DVO2 is the module that outputs to LVDS. Add DVO2-relative definitions
and mt8189_mtk_ddp_main_routes_2 to route DVO2 in the main display path.
Since DSI or LVDS share common output pins, it is not possible to output
DSI and LVDS at the same time. Therefore, DVO2 and DSI are placed in the
same main display path.

Change-Id: I291140f86049ada0ff50b04c509ac9add9a3e3d6
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>