GENIO: media: mediatek: isp_30: seninf: Fix pad link creation
The commit fb2bd86270cd ("media: mc: fix endpoint iteration") uses
src_sd->fwnode instead of dev_fwnode(src_sd->dev) to iterate all
endpoints causing the SENINF to create the source-to-sink link failed.
The reason is that, the SENINF subdev is not registered as an async
subdev and the subdev->fwnode is not set correctly. To address this
issue, manually set the subdev->fwnode to dev_fwnode(dev) when
initializing the v4l2 subdev.

CR-Id: AUTO00774487
Change-Id: Ie4dce0dc3782a7ba611bccf2c05f35f2ee3e2c98
Signed-off-by: Andy.Hsieh <andy.hsieh@mediatek.com>