GENIO: media: mediatek: isp_6s: seninf: Make IRQ resource optional
Previously, the SENINF driver required an IRQ resource to be present,
failing probe if the IRQ was missing. However, the SENINF IRQ may be
optional.

This patch updates the probe logic to use platform_get_irq_optional()
instead of platform_get_irq(), allowing the driver to continue
initialization even if no IRQ resource is available. If the IRQ is
present, it is requested as before; otherwise, the driver logs that no
IRQ resource is available and proceeds.

This improves compatibility with platforms where the SENINF IRQ is not
required.

Change-Id: I22081274573983b550b2f235ae6b8ad0f1c19cd9
Signed-off-by: Andy.Hsieh <andy.hsieh@mediatek.com>