GENIO: drm/mediatek: panel: add drm_mode_set_name to KOE TX18D204VM0BAA LVDS panel
When running modetest on the KOE TX18D204VM0BAA LVDS panel with the
command:

modetest -M mediatek -s 32@45:1920x1080

the command cannot be executed because the mode 1920x1080 is not set
as a valid name. To resolve this, add drm_mode_set_name in the
koe_panel_get_modes function of the KOE TX18D204VM0BAA LVDS panel driver.

Change-Id: I3e60c3548f5767f1965506fb773ebf1a1e87ae01
CR-Id: AUTO00780851
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>